feat(transactions): create a Transaction with a transactions service
The transactions service is a business layer that handles the logic of
the Transactions Domain; e.g. creating, updating and inspecting
transactions. This commit only implements creation logic, and the
Transaction object it uses needs to be fleshed out according to the
needs of the domain.

The business layer is decoupled from persistence logic by using dependency
injection to supply a swappable repository object, which has
responsibility for interactions with a data source. The current
implementation of the repository is very bare bones - an in-memory list.
In the future, the repositroy will access a more complex data source -
a database.