feat(transactions): persist Transactions created by TransactionsService
This feature is tested by showing that when two transactions are created
by the TransactionsSerivice they can subsequently be listed by the
TransactionsService.

The TransactionsService currently utilises an in-memory persistence object,
TransactionsRepository. In the future, this will be refactored to utilise a
database. Changing the persistence mechanism is as simple as changing what
repository object is used, as TransactionsService is loosely coupled to
persistence logic via dependency injection of the
TransactionsRepository.